Move reducer helper functions to separate file
